sakad-backend/requirements.txt
fastapi
uvicorn[standard]
supabase
transformers
torch
torchvision
Pillow
python-dotenv
python-multipart
sentencepiece
protobuf
pydantic-settings
